Update containable.rst
There is an important caveat to using Containable when filtering on a deeper association. In the previous example,
assume you had 3 posts in your database and Daniel had commented on 2 of those posts. The operation
$this->Post->find('all', array('contain' => 'Comment.author = "Daniel"')); would return ALL 3 posts, not
just the 2 posts that Daniel had commented on. It won't return all comments however, just comments by Daniel.::